+++ src/com/kaylerrenslow/armaDialogCreator/control/ControlClass.java
@@ -327,4 +327,8 @@ public class ControlClass {
 		return true;
 	}
 
+	@Override
+	public String toString() {
+		return getClassName();
+	}
 }